Add in provider options for DB Connection Pools
The defaults maintain the current configuration, but this allows users to tune db connections for individual needs.

In particular, even with the latest go libraries, we often encounter "cannot allocate memory" errors when creating yet another DB connection. We have a large number of resources across multiple DBs, so this happens frequently. For our uses, enabling connection pooling is going to save us a lot of headaches and the edge case disabling the connection pooling is guarding against unlikely to be an issue.